requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.1; python_version >= '3.6'
aiosignal==1.2.0; python_version >= '3.6'
async-timeout==4.0.2; python_version >= '3.6'
attrs==22.1.0; python_version >= '3.5'
autopep8==1.7.0
axolpy-lib==1.4.6
base58==2.1.1; python_version >= '3.5'
bitarray==2.6.0
certifi==2022.6.15; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.0; python_version >= '3.6'
cryptography==37.0.4
cython==0.29.32
cytoolz==0.12.0; implementation_name == 'cpython'
eth-abi==2.2.0; python_version >= '3.6' and python_version < '4'
eth-account==0.5.9; python_version >= '3.6' and python_version < '4'
eth-hash[pycryptodome]==0.5.0; python_version >= '3.7' and python_version < '4'
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1; python_version >= '3.6' and python_version < '4'
eth-typing==2.3.0; python_version >= '3.5' and python_version < '4'
eth-utils==1.9.5; python_version >= '3.5' and python_version < '4' and python_full_version != '3.5.2'
frozenlist==1.3.1; python_version >= '3.7'
hexbytes==0.2.3; python_version >= '3.6' and python_version < '4'
idna==3.3; python_version >= '3.5'
iniconfig==1.1.1
ipfshttpclient==0.8.0a2; python_full_version >= '3.6.2' and python_full_version not in '3.7.0, 3.7.1'
jsonschema==4.10.1; python_version >= '3.7'
lru-dict==1.1.8
multiaddr==0.0.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
multidict==6.0.2; python_version >= '3.7'
netaddr==0.8.0
packaging==21.3; python_version >= '3.6'
parsimonious==0.8.1
pluggy==1.0.0; python_version >= '3.6'
protobuf==3.20.1; python_version >= '3.7'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
py-solc-x==1.1.1
pycodestyle==2.9.1; python_version >= '3.6'
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyrsistent==0.18.1; python_version >= '3.7'
pytest==7.1.2; python_version >= '3.7'
pyyaml==6.0; python_version >= '3.6'
requests==2.28.1; python_version >= '3.7' and python_version < '4'
rlp==2.0.1
semantic-version==2.10.0; python_version >= '2.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version >= '3.7'
toolz==0.12.0; python_version >= '3.5'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
varint==1.0.2
web3==5.30.0
websockets==9.1; python_full_version >= '3.6.1'
yarl==1.8.1; python_version >= '3.7'
